hbm-ddr-studio/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pyyaml>=6.0
websockets>=12.0
python-multipart>=0.0.9
apscheduler>=3.10
sqlalchemy>=2.0
